content/canvas/test/test_2d.drawImage.animated.apng.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.drawImage.animated.apng</title>
<!-- Testing: drawImage() of an APNG with no poster frame draws the first frame -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
function isPixel(ctx, x,y, r,g,b,a, pos, colour, d) {
    var pixel = ctx.getImageData(x, y, 1, 1);
    var pr = pixel.data[0],
        pg = pixel.data[1],
        pb = pixel.data[2],
        pa = pixel.data[3];
    ok(r-d <= pr && pr <= r+d &&
       g-d <= pg && pg <= g+d &&
       b-d <= pb && pb <= b+d &&
       a-d <= pa && pa <= a+d,
       "pixel "+pos+" is "+pr+","+pg+","+pb+","+pa+"; expected "+colour+" +/- "+d);
}
function todo_isPixel(ctx, x,y, r,g,b,a, pos, colour, d) {
    var pixel = ctx.getImageData(x, y, 1, 1);
    var pr = pixel.data[0],
        pg = pixel.data[1],
        pb = pixel.data[2],
        pa = pixel.data[3];
    todo(r-d <= pr && pr <= r+d &&
       g-d <= pg && pg <= g+d &&
       b-d <= pb && pb <= b+d &&
       a-d <= pa && pa <= a+d,
       "pixel "+pos+" is "+pr+","+pg+","+pb+","+pa+"; expected "+colour+" +/- "+d);
}
function deferTest() {
    _deferred = true;
}
function wrapFunction(f) {
    return function () {
        f.apply(null, arguments);
        SimpleTest.finish();
    }
}

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

deferTest();
setTimeout(wrapFunction(function () {
    ctx.drawImage(document.getElementById('anim-gr.png'), 0, 0);
    todo_isPixel(ctx, 50,25, 0,255,0,255, "50,25", "0,255,0,255", 2);
}), 500);


});
</script>
<img src="image_anim-gr.png" id="anim-gr.png" class="resource">
